bloggie/requirements.txt
attr==0.3.1
attrs==21.4.0
Django==3.2.13
django_allauth==0.51.0
Pillow==9.1.1
gunicorn==20.1.0
django-crispy-forms ==1.14.0
crispy-forms-semantic-ui>=0.2.0 # added to ease up testing of semantic ui templates

boto3==1.9.96
botocore==1.12.96
certifi==2018.10.15
chardet==3.0.4
dj-database-url==0.5.0
django-heroku==0.3.1
django-storages==1.7.1
docutils==0.14
idna==2.7
jmespath==0.9.3
# psycopg2==2.7.7
python-dateutil==2.8.0
pytz==2018.5
requests==2.19.1
s3transfer==0.2.0
six==1.12.0
urllib3==1.23
whitenoise==6.2.0